#125 review (#129)
* reuse docker pool for metric tests with core

* parse flags on tests

* lint
