+++ client/metric.go
@@ -58,3 +58,18 @@ func (c *Client) AggregatorMetrics(ctx context.Context, aggregatorID string, par
 	path := "/v1/aggregator_metrics/" + url.PathEscape(aggregatorID) + "?" + q.Encode()
 	return out, c.do(ctx, http.MethodGet, path, nil, &out)
 }
+
+// AggregatorPipelinesMetrics get the metrics for a set of pipelineIDs belonging to an aggregator (bulk mode).
+func (c *Client) AggregatorPipelinesMetrics(ctx context.Context, aggregatorID string, params types.PipelinesMetricsParams) (types.PipelinesMetrics, error) {
+	var out types.PipelinesMetrics
+	q := url.Values{
+		"start":    []string{params.Start.String()},
+		"interval": []string{params.Interval.String()},
+	}
+
+	for _, pipelineID := range params.PipelineIDs {
+		q.Add("pipeline_id", pipelineID)
+	}
+
+	return out, c.do(ctx, http.MethodGet, "/v1/aggregators/"+url.PathEscape(aggregatorID)+"/pipelines_metrics?"+q.Encode(), nil, &out)
+}
